requirements.txt
streamlit
langchain
python-dotenv
faiss-cpu
google-generativeai
langchain_google_genai
PyPDF2
pypdf